Working on converting everything to functional components cause why not
