+++ src/components/BaseForm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { withTheme, FormProps } from '@rjsf/core';
-import { Theme as Bootstrap4Theme } from '@rjsf/bootstrap-4';
-import validator from '@rjsf/validator-ajv8';
+import { withTheme, FormProps } from "@rjsf/core";
+import { Theme as Bootstrap4Theme } from "@rjsf/bootstrap-4";
+import validator from "@rjsf/validator-ajv8";
 
 import TagsField from "./TagsField";
 
@@ -9,7 +9,7 @@ const adminFields = { tags: TagsField };
 
 const FormWithTheme = withTheme(Bootstrap4Theme);
 
-const BaseForm = (props:FormProps) => {
+const BaseForm = (props: FormProps) => {
   const { className, ...restProps } = props;
 
   return (
@@ -21,6 +21,6 @@ const BaseForm = (props:FormProps) => {
       fields={adminFields}
     />
   );
-}
+};
 
 export default BaseForm;
