+++ src/components/bucket/BucketAttributes.tsx
@@ -1,3 +1,4 @@
+import React, { useCallback } from "react";
 import type {
   Capabilities,
   BucketState,
@@ -5,8 +6,6 @@ import type {
   SessionState,
   BucketRouteMatch,
 } from "../../types";
-
-import React, { PureComponent } from "react";
 
 import * as BucketActions from "../../actions/bucket";
 import Spinner from "../Spinner";
@@ -29,51 +28,60 @@ export type Props = OwnProps &
     deleteBucket: typeof BucketActions.deleteBucket;
   };
 
-export default class BucketAttributes extends PureComponent<Props> {
-  deleteBucket = (bid: string) => {
-    const { deleteBucket } = this.props;
-    const message = [
-      "This will delete the bucket and all the collections and",
-      "records it contains. Are you sure?",
-    ].join(" ");
-    if (confirm(message)) {
-      deleteBucket(bid);
-    }
-  };
+const BucketAttributes: React.FC<Props> = ({
+  match,
+  session,
+  bucket,
+  capabilities,
+  updateBucket,
+  deleteBucket,
+}) => {
+  const {
+    params: { bid },
+  } = match;
+  const { busy, data: formData } = bucket;
 
-  onSubmit = (formData: BucketData) => {
-    const { match, updateBucket } = this.props;
-    const {
-      params: { bid },
-    } = match;
-    updateBucket(bid, { data: formData });
-  };
+  const handleDeleteBucket = useCallback(
+    (bid: string) => {
+      const message = [
+        "This will delete the bucket and all the collections and",
+        "records it contains. Are you sure?",
+      ].join(" ");
+      if (confirm(message)) {
+        deleteBucket(bid);
+      }
+    },
+    [deleteBucket]
+  );
 
-  render() {
-    const { match, session, capabilities, bucket } = this.props;
-    const {
-      params: { bid },
-    } = match;
-    const { busy, data: formData } = bucket;
-    if (busy) {
-      return <Spinner />;
-    }
-    return (
-      <div>
-        <h1>
-          Edit <b>{bid}</b> bucket attributes
-        </h1>
-        <BucketTabs bid={bid} capabilities={capabilities} selected="attributes">
-          <BucketForm
-            session={session}
-            bid={bid}
-            bucket={bucket}
-            formData={formData}
-            deleteBucket={this.deleteBucket}
-            onSubmit={this.onSubmit}
-          />
-        </BucketTabs>
-      </div>
-    );
+  const handleSubmit = useCallback(
+    (formData: BucketData) => {
+      updateBucket(bid, { data: formData });
+    },
+    [bid, updateBucket]
+  );
+
+  if (busy) {
+    return <Spinner />;
   }
-}
+
+  return (
+    <div>
+      <h1>
+        Edit <b>{bid}</b> bucket attributes
+      </h1>
+      <BucketTabs bid={bid} capabilities={capabilities} selected="attributes">
+        <BucketForm
+          session={session}
+          bid={bid}
+          bucket={bucket}
+          formData={formData}
+          deleteBucket={handleDeleteBucket}
+          onSubmit={handleSubmit}
+        />
+      </BucketTabs>
+    </div>
+  );
+};
+
+export default BucketAttributes;
